Added support for conditional compilation by gcc-xml.
When gcc-xml is executed, we now define a C macro, XDRESS.  This
allows users to use #ifndef statements to hide parts of the code that
we don't want xdress to worry about.

TODO: We should set the value of this macro to the current version
number or string.

TODO: The changes to clang_describe are untested.