Make clang_describe treat ./file.cpp the same as file.cpp
Previously if onlyin contained 'file.cpp', we would skip it since clang
reports the path as './file.cpp'.  This fixes detection of function
argument names in the integration tests.

To make sure this is detected in future, the integration tests now
verify that func1 has the keyword arguments, and that they work in any
order.