Fix integration tests to not unload/reload modules
Unloading and reloading extension modules is not supported
by Python.  This is a problem we run the integration tests
on multiple parsers (gccxml and clang), causing tests to
incorrectly pass or segfault.  Now different parsers build
into different directories, so that the resulting different
modules can be loaded simultaneously.

Slightly modified by Geoffrey from an initial pair of commits
by Anthony, and squashed to avoid intentional breakage in the
middle of the sequence (since the goal was to reproduce test
failures).